+++ emulation/components/support/i8255_ppi.hpp
@@ -23,8 +23,8 @@ namespace Volt
             void Tick() override { }; // Not used here 
             void Shutdown() override;
         
-            uint8_t PortRead(uint8_t port) override;
-            void PortWrite(uint8_t port, uint8_t value) override;
+            uint8_t PortRead(uint16_t port) override;
+            void PortWrite(uint16_t port, uint8_t value) override;
 
             void KeyPress(uint8_t code) override;
